More Coverity warnings
Mostly as a result of 	68f728ab3d84f5dd4b784151e1218b886369d27a
"make ICC profile inclusion optional" in the PNG device.

One was an oversight in the last round of Coverity warnings.

+++ devices/rinkj/rinkj-epson870.c
@@ -944,7 +944,7 @@ rinkj_escp_flush (RinkjEscp *z)
                                          rle,
                                          z->head_bps,
                                          (int)(xsb_out & 0xff),
-                                         (xsb_out >> 8) & 0xff,
+                                         (int)(xsb_out >> 8) & 0xff,
                                          m & 0xff, m >> 8);
       if (status < 0)
 	{
